src/requirements.txt
celery==5.2.6
Django==3.2.4
django-ckeditor==6.1.0
django-environ==0.9.0
django-js-asset==1.2.2
djangorestframework==3.12.4
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.1.1
entrypoints==0.3
gunicorn==20.1.0
httplib2==0.14.0
idna==2.10
importlib-metadata==1.5.0
isort==5.7.0
Mako==1.1.0
MarkupSafe==1.1.1
Pillow==9.3.0
psycopg2-binary==2.9.2

